third_party/blink/web_tests/svg/dynamic-updates/SVGFESpecularLightingElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFESpecularLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS specularElement.in1.baseVal is "SourceAlpha"
PASS specularElement.in1.baseVal is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
